+++ src/_tests/test_1-tiled_dataset.py
@@ -1,5 +1,7 @@
 import numpy as np
+
 from ..tiled_dataset import TiledDataset
+
 
 def test_with_mask_training(client):
     tiled_dataset = TiledDataset(
