+++ src/seg_utils.py
@@ -4,9 +4,10 @@ import numpy as np
 import pandas as pd
 import torch
 from dlsia.core.train_scripts import segmentation_metrics
-from dvclive import Live
 from torch.utils.data import DataLoader, random_split
 from torch.utils.data.dataloader import default_collate
+
+from dvclive import Live
 
 
 def custom_collate(batch):
